Refactor M2D and D2M to Indexed
* Fixes circular dependency on ar.ext.server.ARServer
* Moves MetersToDegrees and DegreesToMeteres
