+++ java/app/ar/app/components/sequentialComposer/OptionDataset.java
@@ -12,7 +12,6 @@ import java.util.Set;
 import java.util.function.Function;
 
 import ar.Glyphset;
-import ar.ext.server.ARServer;
 import ar.glyphsets.MemMapList;
 import ar.glyphsets.SyntheticGlyphset;
 import ar.glyphsets.implicitgeometry.Indexed;
@@ -177,7 +176,7 @@ public final class OptionDataset<G,I> {
 				"US Census Synthetic People", 
 				new File("../data/2010Census_RacePersonPoints.hbin"),
 				//new Indexed.ToPoint(false, 0, 1),
-				new ARServer.MetersToDegrees(0, 1),
+				new Indexed.MetersToDegrees(0, 1),
 				new Indexed.ToValue<Indexed,Character>(2),
 				OptionAggregator.COC_COMP,
 				"(seq(colorkey(cableColors))(catInterpolate(color,clear),.1))",
@@ -204,7 +203,7 @@ public final class OptionDataset<G,I> {
 				"US Census Synthetic People (NY)", 
 				new File("../data/2010Census_RacePersonPoints_NY.hbin"), 
 				//new Indexed.ToPoint(false, 0, 1),
-				new ARServer.MetersToDegrees(0, 1),
+				new Indexed.MetersToDegrees(0, 1),
 				new Indexed.ToValue<Indexed,Character>(2),
 				OptionAggregator.COC_COMP,
 				"(seq(colorkey(cableColors))(catInterpolate(color,clear),.1))",
